Refactor AI content generation services and update AIIdeasViewModel to utilize new service architecture
